backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
volcengine-python-sdk[ark]==4.0.26